fix: отладка функционала

+++ main-service/src/main/java/ru/practicum/ewm/services/category/CategoryServiceImpl.java
@@ -6,11 +6,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.ewm.dto.category.CategoryDto;
 import ru.practicum.ewm.dto.category.CategoryMapper;
+import ru.practicum.ewm.exceptions.ConflictException;
+import ru.practicum.ewm.exceptions.ValidationException;
 import ru.practicum.ewm.models.Category;
 import ru.practicum.ewm.repositories.CategoryRepository;
 import ru.practicum.ewm.exceptions.NotFoundException;
+import ru.practicum.ewm.repositories.EventRepository;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -19,9 +23,11 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto addCategory(CategoryDto dto) {
+        checkUniqueCatName(dto.getName(), null);
         Category category = CategoryMapper.mapDtoToCategory(dto);
         Category savedCategory = categoryRepository.save(category);
         log.info("Добавлена новая категория: {}", savedCategory);
@@ -30,9 +36,15 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto updateCategory(CategoryDto dto) {
+        if (dto.getId() == null) {
+            throw new ValidationException("Не определен id категории");
+        }
         if (!categoryRepository.existsById(dto.getId())) {
             throw new NotFoundException(String.format("Category with id=%d was not found", dto.getId()));
         }
+
+        checkUniqueCatName(dto.getName(), dto.getId());
+
         Category category = CategoryMapper.mapDtoToCategory(dto);
         Category savedCategory = categoryRepository.save(category);
         log.info("Категория изменена: {}", savedCategory);
@@ -44,6 +56,10 @@ public class CategoryServiceImpl implements CategoryService {
         if (!categoryRepository.existsById(catId)) {
             throw new NotFoundException(String.format("Category with id=%d was not found", catId));
         }
+        if (eventRepository.existsByCatId(catId)) {
+            throw new ConflictException("Нельзя удалить категорию, по которой существуют события");
+        }
+
         categoryRepository.deleteById(catId);
         log.info("Категория с id={} удалена", catId);
     }
@@ -61,6 +77,15 @@ public class CategoryServiceImpl implements CategoryService {
         Category category = categoryRepository.findById(catId).orElseThrow(() ->
                 new NotFoundException("Категория не найдена"));
         return CategoryMapper.mapCategoryToDto(category);
+    }
+
+    private void checkUniqueCatName(String name, Long id) {
+        Category category = categoryRepository.findByName(name);
+        if (category != null) {
+            if (id == null || !id.equals(category.getId())) {
+                throw new ConflictException("Категория с таким названием уже существует");
+            }
+        }
     }
 
 }
